requirements.txt
pyannote.audio==3.1.1
openai-whisper
numpy<2
torch==2.2.0
torchaudio==2.2.0
ffmpeg-python==0.2.0
soundfile
huggingface_hub==0.16.4
openai>=1.0.0
anthropic>=0.18.0
pyyaml>=6.0
